Move validate_role_service to layman_start_util.py
